CLI: add structure filters to `analyse cif-clean`
After the workchains had all been run, it was found that a number of
incorrect CIFs went unnoticed. These were found in a post processing run
and marked with various tags in the extras. These CIFs should be
excluded in the analysis as being valid CIFs. The `analyse cif-clean`
command has a new option that allows specifying tags that should filter
out `CifData` nodes that have those tags in their extras.